+++ js/util.js
@@ -1,11 +1,4 @@
-const ALERT_SHOW_TIME = 5000;
-//генераор случайных чисел
-const getRandomInteger = (a, b) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-};
+import {ALERT_SHOW_TIME, DEBOUNCE_TIME } from './data.js';
 
 const isEscapeKey = (evt) => evt.key === 'Escape';
 
@@ -17,7 +10,7 @@ const showAlert = (message) => {
   alertContainer.style.top = '0';
   alertContainer.style.right = '0';
   alertContainer.style.padding = '10px 3px';
-  alertContainer.style.fontSize = '25px';
+  alertContainer.style.fontSize = '22px';
   alertContainer.style.textAlign = 'center';
   alertContainer.style.borderColor = '#e90000';
   alertContainer.style.borderRadius = '3px';
@@ -33,7 +26,7 @@ const showAlert = (message) => {
   }, ALERT_SHOW_TIME);
 };
 
-const debounce = (callback, timeoutDelay = 500) => {
+const debounce = (callback, timeoutDelay = DEBOUNCE_TIME) => {
   let timeoutId;
 
   return (...rest) => {
@@ -42,4 +35,4 @@ const debounce = (callback, timeoutDelay = 500) => {
   };
 };
 
-export {getRandomInteger, isEscapeKey,showAlert,debounce};
+export {isEscapeKey,showAlert,debounce};
